ナビゲータ
・barのファイル作成
・ナビゲータの修正
