+++ lib/history.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
+import 'package:team/bar.dart';
 import 'date.dart';
 import 'main.dart';
 import 'calo.dart';
@@ -52,7 +53,8 @@ class HistoryPage extends State<HistoryState> {
 
   @override
   Widget build(BuildContext context) {
-    final sortedDates = _historyData.keys.toList()..sort((a, b) => b.compareTo(a));
+    final sortedDates = _historyData.keys.toList()
+      ..sort((a, b) => b.compareTo(a));
     return Scaffold(
       appBar: AppBar(
         title: const Text('歩数の履歴'),
@@ -62,57 +64,80 @@ class HistoryPage extends State<HistoryState> {
       body: _isLoading
           ? const Center(child: CircularProgressIndicator())
           : sortedDates.isEmpty
-              ? const Center(child: Text('履歴データがありません'))
-              : ListView.builder(
-                  padding: const EdgeInsets.all(8.0),
-                  itemCount: sortedDates.length,
-                  itemBuilder: (context, index) {
-                    final dateString = sortedDates[index];
-                    final steps = _historyData[dateString]!;
+          ? const Center(child: Text('履歴データがありません'))
+          : ListView.builder(
+              padding: const EdgeInsets.all(8.0),
+              itemCount: sortedDates.length,
+              itemBuilder: (context, index) {
+                final dateString = sortedDates[index];
+                final steps = _historyData[dateString]!;
 
-                    //カロリー
-                    final calClass = Calorie(steps);
-                    final consumeCal = calClass.Kcal;
-                    final consumeFat = calClass.fat;
-                    
-                    return Card(
-                      child: ListTile(
-                        leading: const Icon(Icons.directions_walk, color: Colors.blue),
+                //カロリー
+                final calClass = Calorie(steps);
+                final consumeCal = calClass.Kcal;
+                final consumeFat = calClass.fat;
 
-                        title: Text(
-                          _formatDisplayDate(dateString),
-                          style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
-                        ),
-                        subtitle: Text(
-                          '${consumeCal.toStringAsFixed(1)} Kcal  /  ${consumeFat.toStringAsFixed(1)} g',
-                          style: TextStyle(color: Colors.grey.shade600),
-                        ),
-                        trailing: Text(
-                          '${formatter.format(steps)}歩',
-                          textAlign: TextAlign.right,
-                          style: const TextStyle(
-                            fontSize: 18,
-                            fontWeight: FontWeight.w600,
-                            color: Colors.blue,
-                          ),
-                        ),
+                return Card(
+                  child: ListTile(
+                    leading: const Icon(
+                      Icons.directions_walk,
+                      color: Colors.blue,
+                    ),
+
+                    title: Text(
+                      _formatDisplayDate(dateString),
+                      style: const TextStyle(
+                        fontWeight: FontWeight.bold,
+                        fontSize: 16,
                       ),
-                    );
-                  },
-                ),
+                    ),
+                    subtitle: Text(
+                      '${consumeCal.toStringAsFixed(1)} Kcal  /  ${consumeFat.toStringAsFixed(1)} g',
+                      style: TextStyle(color: Colors.grey.shade600),
+                    ),
+                    trailing: Text(
+                      '${formatter.format(steps)}歩',
+                      textAlign: TextAlign.right,
+                      style: const TextStyle(
+                        fontSize: 18,
+                        fontWeight: FontWeight.w600,
+                        color: Colors.blue,
+                      ),
+                    ),
+                  ),
+                );
+              },
+            ),
       bottomNavigationBar: BottomAppBar(
         color: Colors.blue.shade100,
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceAround,
           children: [
-
             IconButton(icon: const Icon(Icons.menu), onPressed: () {}),
-            IconButton(icon: const Icon(Icons.home), onPressed: () {
-              Navigator.push(context, MaterialPageRoute(builder: (context) => PedometerScreen()));
-            }),
-            IconButton(icon: const Icon(Icons.bar_chart), onPressed: () {
-              Navigator.push(context, MaterialPageRoute(builder: (context) => NextState(stepCount: 0,barGroups: [],)));
-            }),
+            IconButton(
+              icon: const Icon(Icons.home),
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => PedometerScreen()),
+                );
+              },
+            ),
+            IconButton(
+              icon: const Icon(Icons.bar_chart),
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => NextState(
+                      stepCount: 0,
+                      dateManager: widget.dateManager,
+                      barGroups: [],
+                    ),
+                  ),
+                );
+              },
+            ),
           ],
         ),
       ),
