+++ lib/main.dart
@@ -2,12 +2,11 @@ import 'dart:async';
 import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:sensors_plus/sensors_plus.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'calo.dart';
 import 'date.dart';
 import 'package:intl/intl.dart';
 import 'history.dart';
 import 'package:fl_chart/fl_chart.dart';
+import 'package:team/bar.dart';
 
 final formatter = NumberFormat("#,###");
 // --- アプリケーションのエントリーポイント ---
@@ -51,12 +50,15 @@ class _PedometerScreenState extends State<PedometerScreen> {
   bool _isPeak = false;
 
   List<BarChartGroupData> barGroups = [
-          BarChartGroupData(x: 1, barRods: [
-          BarChartRodData(toY: 30.toDouble(), width: 15, color: Colors.green),
-        ]),
-          ];//日付と歩数  
+    BarChartGroupData(
+      x: 1,
+      barRods: [
+        BarChartRodData(toY: 30.toDouble(), width: 15, color: Colors.green),
+      ],
+    ),
+  ]; //日付と歩数
   late int bord;
-  int ind=1;
+  int ind = 1;
 
   @override
   void initState() {
@@ -68,14 +70,14 @@ class _PedometerScreenState extends State<PedometerScreen> {
     // 今日の日付を取得し、その日付の歩数を読み込む
     _currentDate = _dateManager.getTodaydate();
     final savedSteps = await _dateManager.loadStep(_currentDate);
-    final savedList = await _dateManager.loadList(_currentDate,barGroups);
-    final savedind  = await _dateManager.loadind(_currentDate);
+    final savedList = await _dateManager.loadList(_currentDate, barGroups);
+    final savedind = await _dateManager.loadind(_currentDate);
     if (mounted) {
       setState(() {
         _stepCount = savedSteps;
-        barGroups=savedList;
-        bord=_stepCount+10;
-        ind =savedind;
+        barGroups = savedList;
+        bord = _stepCount + 10;
+        ind = savedind;
       });
     }
     _startListening(); // センサーの監視を開始
@@ -91,12 +93,19 @@ class _PedometerScreenState extends State<PedometerScreen> {
         // 日付が変わっていたら、前日(_currentDateString)の歩数(_stepCount)を保存
         await _dateManager.saveStep(_currentDate, _stepCount);
         barGroups.add(
-          BarChartGroupData(x: ind, barRods: [
-          BarChartRodData(toY: _stepCount.toDouble(), width: 15, color: Colors.green),
-        ]),
-          );
+          BarChartGroupData(
+            x: ind,
+            barRods: [
+              BarChartRodData(
+                toY: _stepCount.toDouble(),
+                width: 15,
+                color: Colors.green,
+              ),
+            ],
+          ),
+        );
         ind++;
-        await _dateManager.saveStepList(_currentDate,barGroups);
+        await _dateManager.saveStepList(_currentDate, barGroups);
         await _dateManager.saveStep(_currentDate, ind);
 
         // 新しい日のためにリセット
@@ -200,12 +209,13 @@ class _PedometerScreenState extends State<PedometerScreen> {
 
           children: [
             IconButton(
-              icon: const Icon(Icons.menu), 
+              icon: const Icon(Icons.menu),
               onPressed: () {
                 Navigator.push(
                   context,
                   MaterialPageRoute(
-                    builder: (context) => HistoryState(dateManager: _dateManager),
+                    builder: (context) =>
+                        HistoryState(dateManager: _dateManager),
                   ),
                 );
               },
@@ -219,105 +229,13 @@ class _PedometerScreenState extends State<PedometerScreen> {
                   MaterialPageRoute(
                     builder: (context) => NextState(
                       stepCount: _stepCount,
-                      barGroups :barGroups,
-                      ),
+                      dateManager: _dateManager,
+                      barGroups: barGroups,
+                    ),
                   ),
                 );
               },
             ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-double distance = 0.0;
-
-class NextState extends StatefulWidget {
-  final int stepCount;
-  final List<BarChartGroupData> barGroups;
-  const NextState({super.key, required this.stepCount,required this.barGroups});
-
-  @override
-  State<NextState> createState() => NextPage();
-}
-
-class NextPage extends State<NextState> {
-  final TextEditingController textController = TextEditingController();
-  late int _localStepCount;
-  late double consume_cal;
-  late double consumeFat;
-  late List<BarChartGroupData> _barGroups;
-  void calo() {
-    setState(() {
-      final calClass = Calorie(_localStepCount);
-      consume_cal = calClass.Kcal;
-      consumeFat = calClass.fat;
-    });
-  }
-
-  @override
-  initState() {
-    super.initState();
-    _localStepCount = widget.stepCount;
-    _barGroups =widget.barGroups;
-    calo();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('次のページ'),
-        automaticallyImplyLeading: false,
-      ),
-      body: Center(
-         child :Column(
-        children: [Text(
-          '消費カロリー${consume_cal.toStringAsFixed(2)}Kcal\n脂肪燃焼量${consumeFat.toStringAsFixed(2)}g',
-          style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
-        ),
-         const SizedBox(height: 32), // スペーサー
-        SingleChildScrollView( // グラフ全体を横スクロールさせる
-                scrollDirection: Axis.horizontal,
-                child: SizedBox(
-                  // グラフの幅を動的に計算
-                  width: 300+_barGroups.length * 50.0, // 各棒の幅(15) + グループ間のスペース(10) + 余白
-                  height: 500, // チャートに固定の高さを与える
-                  child: BarChart(
-                    BarChartData(
-                      borderData: FlBorderData(
-                        border: const Border(
-                          top: BorderSide.none,
-                          right: BorderSide.none,
-                          left: BorderSide(width: 1),
-                          bottom: BorderSide(width: 1),
-                        ),
-                      ),
-                      groupsSpace: 10,
-                      barGroups: _barGroups, // すべてのデータをまとめて渡す
-                    ),
-                  ),
-                ),
-              ),
-        ]
-      ),
-      ),
-      bottomNavigationBar: BottomAppBar(
-        color: Colors.blue.shade100,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-
-          children: [
-            IconButton(icon: const Icon(Icons.menu), onPressed: () {}),
-            IconButton(
-              icon: const Icon(Icons.home),
-              onPressed: () {
-                Navigator.pop(context);
-              },
-            ),
-            IconButton(icon: const Icon(Icons.bar_chart), onPressed: () {}),
           ],
         ),
       ),
